backend/python-streaming/requirements.txt
backports.csv==1.0.7
beautifulsoup4==4.9.3
better-profanity==0.7.0
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
get-all-tickers==1.7
idna==2.10
kafka-python==2.0.2
numpy==1.20.1
pandas==1.2.2
praw==7.1.4
prawcore==1.5.0
pymongo==3.11.3
python-dateutil==2.8.1
pytimeparse==1.1.8
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
tablib==3.0.0
twelvedata==1.1.8
update-checker==0.18.0
urllib3==1.26.3
websocket-client==0.57.0
Yahoo-ticker-downloader==3.0.1
